Disable waiting for timeline tab list to be populated before removing the splash screen
Because of process death, you can restart the app and come back in another screen than the default one, which require for each screen to manually remove the splash screen.

As this doesn't scale up and can lead to a completely frozen state, I prefer to disable it for now.

It also prevented the article screen to rotate, creating a full black screen.